moved file upload to custom hook
