+++ src/components/right-panel/text-box/fileSend.tsx
@@ -1,12 +1,9 @@
-import React, { useRef, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import { toast } from 'react-toastify';
 import { Room } from 'livekit-client';
 import { useTranslation } from 'react-i18next';
-import Resumable from 'resumablejs';
-import ResumableFile = Resumable.ResumableFile;
 
 import { store } from '../../../store';
-import { ISession } from '../../../store/slices/interfaces/session';
 import {
   isSocketConnected,
   sendWebsocketMessage,
@@ -16,6 +13,7 @@ import {
   IChatMsg,
   IDataMessage,
 } from '../../../store/slices/interfaces/dataMessages';
+import useResumableFilesUpload from '../../../helpers/hooks/useResumableFilesUpload';
 
 interface IFileSendProps {
   isChatServiceReady: boolean;
@@ -29,14 +27,33 @@ const FileSend = ({
   currentRoom,
 }: IFileSendProps) => {
   const inputFile = useRef<HTMLInputElement>(null);
-  const toastId = React.useRef<string>(null);
-  const [isUploading, setIsUploading] = useState<boolean>(false);
   const { t } = useTranslation();
+  const [files, setFiles] = useState<Array<File>>();
+
   const chat_features =
     store.getState().session.currentRoom.metadata?.room_features.chat_features;
   const accept =
     chat_features?.allowed_file_types?.map((type) => '.' + type).join(',') ??
     '*';
+  const maxFileSize = chat_features?.max_file_size
+    ? chat_features?.max_file_size
+    : undefined;
+
+  const { isUploading, result } = useResumableFilesUpload({
+    allowedFileTypes: chat_features?.allowed_file_types ?? [],
+    maxFileSize,
+    files,
+  });
+
+  useEffect(() => {
+    if (result && result.filePath && result.fileName) {
+      publishToChat(result.filePath, result.fileName);
+      toast(t('right-panel.file-upload-success'), {
+        type: toast.TYPE.SUCCESS,
+      });
+    }
+    //eslint-disable-next-line
+  }, [result]);
 
   const openFileBrowser = () => {
     if (!isUploading) {
@@ -49,109 +66,7 @@ const FileSend = ({
     if (!files.length) {
       return;
     }
-    const session = store.getState().session;
-    sendFile(session, files);
-  };
-
-  const sendFile = (session: ISession, files: Array<File>) => {
-    let fileName = '';
-
-    const r = new Resumable({
-      target: (window as any).PLUG_N_MEET_SERVER_URL + '/api/fileUpload',
-      uploadMethod: 'POST',
-      query: {
-        sid: session.currentRoom.sid,
-        roomId: session.currentRoom.room_id,
-        userId: session.currenUser?.userId,
-        resumable: true,
-      },
-      headers: {
-        Authorization: session.token,
-      },
-      fileType: chat_features?.allowed_file_types ?? [],
-      fileTypeErrorCallback(file) {
-        toast(t('notifications.file-type-not-allow', { filetype: file.type }), {
-          type: toast.TYPE.ERROR,
-        });
-      },
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      maxFileSize: chat_features?.max_file_size
-        ? chat_features?.max_file_size * 1000000
-        : undefined,
-      maxFileSizeErrorCallback() {
-        toast(t('notifications.max-file-size-exceeds'), {
-          type: toast.TYPE.ERROR,
-        });
-      },
-    });
-
-    r.on('fileAdded', function (file) {
-      fileName = file.fileName;
-      if (!r.isUploading()) {
-        setIsUploading(true);
-        r.upload();
-      }
-    });
-
-    r.on('fileSuccess', function (file: ResumableFile, message: string) {
-      const res = JSON.parse(message);
-      setIsUploading(false);
-
-      setTimeout(() => {
-        toast.dismiss(toastId.current ?? '');
-      }, 300);
-
-      if (res.status) {
-        publishToChat(res.filePath, res.fileName);
-        toast(t('right-panel.file-upload-success'), {
-          type: toast.TYPE.SUCCESS,
-        });
-      }
-    });
-
-    r.on('fileError', function (file, message) {
-      setIsUploading(false);
-
-      setTimeout(() => {
-        toast.dismiss(toastId.current ?? '');
-      }, 300);
-
-      try {
-        const res = JSON.parse(message);
-        toast(t(res.msg), {
-          type: toast.TYPE.ERROR,
-        });
-      } catch (e) {
-        toast(t('right-panel.file-upload-default-error'), {
-          type: toast.TYPE.ERROR,
-        });
-      }
-    });
-
-    r.on('uploadStart', function () {
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      toastId.current = toast(
-        t('right-panel.uploading-file', {
-          fileName,
-        }),
-        {
-          closeButton: false,
-          progress: 0,
-        },
-      );
-    });
-
-    r.on('fileProgress', function (file) {
-      const progress = file.progress(false);
-
-      toast.update(toastId.current ?? '', {
-        progress: Number(progress),
-      });
-    });
-
-    r.addFiles(files);
+    setFiles([...files]);
   };
 
   const publishToChat = (filePath: string, fileName: string) => {
